arreglar warnings

+++ Kernel/include/registers.h
@@ -1,4 +1,4 @@
 
 void saveReg(uint64_t * stackFrame);
 
-uint16_t * getVec();
+uint64_t * getVec();
