+++ Kernel/registers.c
@@ -16,6 +16,6 @@ void saveReg(uint64_t * stackFrame) {
     }
 }
 
-uint16_t * getVec() {
+uint64_t * getVec() {
     return memory;
 }
